fix: avoid signed left-shift overflow in CRC32_CKSUM_UpdateUsingFormula (#9)
*byte is promoted to int before the <<24 shift, which overflows signed
int range (UB) for byte values >= 0x80. Cast to uint32_t first.

+++ src/crc32.c
@@ -65,7 +65,7 @@ uint32_t CRC32_CKSUM_UpdateUsingFormula(void const *data, size_t data_size, uint
 
   while(data_size--) {
     // with help from https://github.com/Michaelangel007/crc32
-    crc = crc ^ (*byte++ << 24);
+    crc = crc ^ ((uint32_t)*byte++ << 24);
     for(int bit = 0; bit < 8; bit++) {
       if(crc & (1L << 31)) {
         crc = (crc << 1) ^ 0x04C11DB7;
